backnd-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.6.1
email-validator==2.2.0